refactor(mentor): 파일명에 빈칸이 들어가 있는 오류 수정(BannerImage.jsx)
