+++ app/(root)/page.jsx
@@ -1,6 +1,6 @@
 import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
-import BannerImage from '@/components/home/\bBannerImage';
+import BannerImage from '@/components/home/BannerImage';
 import HomeMain from '@/components/home/HomeMain';
 
 export default function HomePage() {
